backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pyodbc>=5.0.0
mysql-connector-python>=8.4.0
psycopg2-binary>=2.9.9
anthropic>=0.25.0
openai>=1.30.0
google-generativeai>=0.7.0
python-multipart>=0.0.9
websockets>=12.0
pydantic>=2.6.0
